requirements.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
Django==3.2.13
idna==3.4
Pillow==9.2.0
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
vk-api==11.9.9
